fix to flag piezo movement and adding test for it
