+++ PYME/Analysis/points/piezo_movement_correction.py
@@ -27,20 +27,15 @@ def flag_piezo_movement(data_source, events, fps=None):
             # ProtocolFocus description is 'frame#, position'
             f_frame, f_position = str(event['EventDescr']).split(',')
             focus_frames.append(float(f_frame))
-            # focus_positions.append(float(f_position))
             focus_times.append(event['Time'])
 
         if str(event['EventName']) == 'PiezoOnTarget':
-            # PiezoOnTarget description is the position
-            # ontarget_positions.append(float(event['EventDescr']))
             ontarget_times.append(float(event['Time']))
 
     # convert to arrays
-    focus_frames = np.asarray(focus_frames)
+    focus_frames = np.asarray(focus_frames, dtype=int)
     focus_times = np.asarray(focus_times)
-    # focus_positions = np.asarray(focus_positions)
     ontarget_times = np.asarray(ontarget_times)
-    # ontarget_positions = np.asarray(ontarget_positions)
 
     I_time_focus = np.argsort(focus_times)
 
@@ -66,7 +61,7 @@ def flag_piezo_movement(data_source, events, fps=None):
     for frame in focus_frames:
         # find the next on-target
         valid_ontargets = ontarget_frames > frame
-        ot_inds = np.argsort(focus_frames - frame)
+        ot_inds = np.argsort(ontarget_frames - frame)
         try:
             ot_ind = ot_inds[valid_ontargets][0]
             ontarget_frame = ontarget_frames[ot_ind]
@@ -88,25 +83,3 @@ def correct_target_positions(data_source, events):
     :return:
     """
     return
-
-
-TEST_EVENTS = np.array([#(b'PiezoOnTarget', 0, b'48.262'),
-       ('ProtocolFocus', 0, '0, 49.988'),
-       ('PiezoOnTarget', 1 * 0.00125, '48.307'),
-       ('ProtocolTask', 1 * 0.00125, '1, DisableLock, '),
-       ('PiezoOffsetUpdate', 2 * 0.00125, '-1.6720'),
-       ('ProtocolFocus', 801 * 0.00125, '801, 51.188'),
-       ('PiezoOnTarget', 850 * 0.00125, '49.489'),
-       ('ProtocolFocus', 1601 * 0.00125, '1601, 52.388'),
-       ('PiezoOnTarget', 1650 * 0.00125, '50.705'),
-       ('ProtocolFocus', 2401 * 0.00125, '2401, 53.588'),
-       ('ProtocolTask', 1.58222556e+09, 'EnableLock, '),
-       ('ProtocolTask', 1.58222556e+09, 'LaunchAnalysis, ')],
-      dtype=[('EventName', 'U32'), ('Time', '<f8'), ('EventDescr', 'U256')])
-
-TEST_DATA_SOURCE = np.arange(2500).astype([('t', '<i4')])
-
-def test_flag_piezo_movement():
-    moving = flag_piezo_movement(TEST_DATA_SOURCE, TEST_EVENTS)
-
-test_flag_piezo_movement()
